refactor: name of secondary screen comps
changed to match titles
