+++ apps/atlas/src/components/panels/RealmDetails/Resources.tsx
